cms_example/requirements.txt
django-page-cms
django-filebrowser-no-grappelli
django-mptt
django-authority

